# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Student Leader Cabinet API
  version: 0.6.0
  description: >
    REST API личного кабинета студенческого лидера. Все ответы обёрнуты в envelope
    {data, meta, request_id} либо {error, request_id} (SITE.md §20, §50).
servers:
  - url: https://eazytech.ru/api/v1
    description: production
  - url: http://localhost:8080/api/v1
    description: local
paths:
...
  /participant/lectures:
    get:
      tags: [Attendance]
      summary: Лекции и посещения текущего участника
      operationId: participantLectures
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
...
  /participant/orders:
    get:
      tags: [Merch]
      summary: Заказы текущего участника
      operationId: participantOrders
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Merch]
      summary: Зарезервировать товары и баллы
      operationId: participantReserveOrder
      security: [{ participantCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveOrderInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '201': { $ref: '#/components/responses/Created' }
        default: { $ref: '#/components/responses/Error' }
  /participant/orders/{orderId}:
    parameters: [{ $ref: '#/components/parameters/OrderId' }]
    get:
      tags: [Merch]
      summary: Собственный заказ
      operationId: participantOrder
      security: [{ participantCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 200 }
...
  responses:
    Success:
      description: Успешный ответ в общем envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
    Created:
      description: Ресурс создан; идемпотентный replay может вернуть 200
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Error:
      description: Ошибка в общем envelope; HTTP status и error.code зависят от причины
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data: {}
        meta:
          type: object
          nullable: true
        request_id:
          type: string
          format: uuid
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_REFRESH_REUSED
                - AUTH_ACCOUNT_BLOCKED
                - AUTH_PASSWORD_CHANGE_REQUIRED
                - FORBIDDEN
                - CONTEST_ACCESS_DENIED
                - RESOURCE_NOT_FOUND
                - SUBMISSION_LOCKED
                - SUBMISSION_DEADLINE_PASSED
                - SUBMISSION_VERSION_CONFLICT
                - SUBMISSION_ALREADY_SUBMITTED
                - UPLOAD_NOT_COMPLETED
                - FILE_TYPE_NOT_ALLOWED
                - FILE_TOO_LARGE
                - IDEMPOTENCY_CONFLICT
                - RATE_LIMIT_EXCEEDED
                - TELEGRAM_DELIVERY_FAILED
                - NOT_FOUND
                - PARTICIPANT_IDENTIFIER_TAKEN
                - PARTICIPANT_IDENTITY_AMBIGUOUS
                - PARTICIPANT_AUTH_FAILED
                - PARTICIPANT_SESSION_EXPIRED
                - LECTURE_NOT_FOUND
                - INVALID_LECTURE_TRANSITION
                - LECTURE_HAS_ATTENDANCE
                - ATTENDANCE_CLOSED
                - QR_INVALID
                - QR_EXPIRED
                - QR_REPLAYED
                - PARTICIPANT_UNAVAILABLE
                - TASK_NOT_FOUND
                - INVALID_TASK_TRANSITION
                - TASK_SUBMISSION_CLOSED
                - STORAGE_UNAVAILABLE
                - MERCH_NOT_FOUND
                - INVALID_MERCH_TRANSITION
                - INSUFFICIENT_STOCK
                - INSUFFICIENT_POINTS
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
              nullable: true
        request_id:
          type: string
...
    ReserveOrderInput:
      type: object
      required: [items]
      properties:
        idempotency_key: { type: string, minLength: 8, maxLength: 200 }
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    participantCookie:
      type: apiKey
      in: cookie
      name: slc_participant_session